requirements.txt
azure-cognitiveservices-vision-customvision==3.1.0
Flask==2.0.2
matplotlib==3.5.1
msrest==0.6.21
numpy==1.22.1
Pillow==9.1.1
python-dotenv==0.20.0
